requirements.txt
absl-py
tensorflow>=1.11.0
numpy>=1.16.2
Pillow
opencv-python
transformers
tqdm
torch
matplotlib